refactor : 메서드 위치 조정
내가 쓴 댓글 조회, 좋아요한 리뷰 조회 등을 member domain에 작성했는데,
이를 각각 reply, review 도메인으로 위치 이동
